MAINT: special: refactor `ufunc.h` and move to `xsf` (#21472)
* refactoring ufunc.h in special

* refactored out relevant parts of gufunc.cpp

* moved ufunc.h to xsf/numpy.h

* Add "xsf/numpy.h" to list of Python-including headers

---------
